// src/app/models/agendamento.ts
export class Agendamento {
    nome_medico: string = "";
    crm: number = 0;
    telefone: string = "";
    prestador: string = "";
    data: string = "";
    hora: string = "";
    user: string = "";
    atendimento: string = "";
    observacao: string = "";
    procedimento: string = "";
    email: string = "";
    encaixe: boolean = false;
}